[RISC-V] Disable default attribute mix warnings
Display attribute mix warnings only when explicitly setting the
-Wattribute-mix option.

Projects currently building with -Werror enabled are failing due to the
default-enabled attribute mix warnings.

It is acceptable to disable these attribute mix warnings because attribute
mixes that can lead to ABI incompatibility issues are treated as errors by
the linker.

Signed-off-by: Ana Pazos <apazos@quicinc.com>