support omagic
Purpose:

--omagic is primarily used when you want to generate an executable with a
specific memory layout and characteristics, particularly when memory space is
limited or when demand paging is not desired.

Key features:

No page alignment: The data segment is not aligned to page boundaries, which can
lead to more compact executables.

Readable and writable sections: Both the text and data sections are marked as
readable and writable, allowing for modifications during execution.

Fixes (#98)

Change-Id: I46480d959c41c175a4898c89ec0ac286fde54bcb
Signed-off-by: Shankar Easwaran <seaswara@quicinc.com>