Add symbols referred by shared libs to dynamic symbol table
This commit fixes the missing symbol from dynamic symbol table issue
when a symbol was referred by dynamic symbol table but was resolved to a
symbol by a reloctable object that was present after the shared library
in the link command.

The issue is fixed by properly propagating exportToDyn property across
symbols during symbol resolution.

Closes #106

Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>