Annotate autoloaded plugins in the text map-file
This commit adds an annotation for autoloaded plugins in the text
map-file. The main motivation for this map-file addition is to be able
to easily tell which plugins were loaded automatically by the linker and
which were loaded by the user.

This commit also adds a dummy HelloWorldPlugin. This serves two purpose:
it provides a skeleton linker plugin for the users to start with and to
provide a plugin for testing the autoload annotation in the map-file.

Closes #136

Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>