[RISCV] Implement Relaxations for QC.E.LI/QC.LI
These relaxations are like LUI relaxation, so they are done in the same
phase.

The relaxations are:
- `QC.E.LI` to `C.LUI` or `QC.E.LI` or `ADDI` (gp-relative)
- `QC.LI` to `C.LUI`

These respect the existing command-line switches which prevent these
being applied in certain cases.

These relaxations are described in
http://github.com/quic/riscv-elf-psabi-quic-extensions/releases/latest .
The C.LUI relaxation is described in
quic/riscv-elf-psabi-quic-extensions#9 and will be merged soon.

This continues the implementation of #147.

Signed-off-by: Sam Elliott <quic_aelliott@quicinc.com>