Fix incorrect evaluation of ternary expression during GC
This commit fixes incorrect evaluation of ternary expressions during
garbage-collection. It was causing a linker assertion failure because
of using an expression value before it was computed.

Previously, before we made expression's value std::nullopt by
default (#212), the same incorrect evaluation caused incorrect
garbage-collection of the required section when a ternary expression
containing a symbol from the section is used to initialize a useful
linker script symbol.

This commit fixes both these issues.

Closes #266

Signed-off-by: Parth Arora <partaror@qti.qualcomm.com>